UI phieu phat

+++ backend2/library/src/main/java/com/library_web/library/model/BorrowCard.java
@@ -119,7 +119,7 @@ public class BorrowCard {
   @Column(name = "status", nullable = false)
   private String status;
 
-  private long soNgayTre;
+  private Integer soNgayTre;
 
   public enum Status {
     REQUESTED("Đã yêu cầu"),
